# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Game Collector App
# during dev, should point to your local machine
#host: localhost:10010
host: advance-path-222122.appspot.com
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /operadores:
    x-swagger-router-controller: drivers.controller
    get:
      description: get the operadores list
      operationId: getDrivers
      #parameters:
      #  - name: name
      #    in: query
      #    type: string
      #    required: false
      #  - name: sort
      #    in: query
      #    type: string
      #    required: false
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetOperadoresListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new operador to the db
      operationId: createOperador
      parameters:
        - in: body
          name: operador
          description: The operador to create.
          schema:
            $ref: "#/definitions/Operador"
          required: true
 
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetOperadorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /viajes:
    x-swagger-router-controller: viajes.controller

    get:
      tags:
      - viajes
      description: "Obtiene la lista de viajes"
      operationId: "getOrders"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GetViajeListResponse"
        404:
          description: "Viajes not found"
    post:
      tags:
      - viajes
      description: "inserta un nuevo viaje"
      operationId: "addViaje"
      produces:
      - "application/json"
      parameters:
        - in: body

          name: myorder
          description: Viaje object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/Order'

          #name: orderdet
          #description: Viaje object that needs to be added
          #required: true
          #schema:
          #  $ref: '#/definitions/GetViajeDetailListResponse'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Viaje'
        '400':
          description: Invalid Order
  '/viajes/{id}':
    x-swagger-router-controller: viajes.controller

    get:
      tags:
        - viajes
      summary: Find Viaje by ID
      description: Returns a single Viaje
      operationId: getOrderbyId
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of viaje to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Viaje'
        '400':
          description: Invalid ID supplied
        '404':
          description: Viaje not found
  '/viajes_detail/{id}':
    x-swagger-router-controller: viajes.controller

    get:
      tags:
        - viajes
      summary: Find Viaje detail by ID
      description: Returns a list of Viaje Detail
      operationId: getOrderDetailbyId
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of viaje detail to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/GetViajeDetailListResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Viaje not found
...
  /rutas:
    x-swagger-router-controller: routes.controller

    get:
      tags:
      - rutas
      description: "Obtiene la lista de rutas"
      operationId: "getRoutes"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GetRouteListResponse"
        404:
          description: "Routes not found"
  '/rutas_detail/{id}':
    x-swagger-router-controller: routes.controller

    get:
      tags:
        - rutas
      summary: Find dolly by ID
      description: Returns a single route
      operationId: getRouteDetailById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of route to return
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/GetRouteDetailListResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Route not found
...
definitions:
...
  Operador:
    type: object
    properties:
      codigo:
        type: integer
      nombre:
        type: string
      apellidos:
        type: string
      id_flota:
        type: string
      tag:
        type: string
      email:
        type: string
      loginmanual:
        type: string

    example:
      codigo: 100
      nombre: 'adolfo'
      apellidos: 'centeno tellez'
      id_flota: 'hesa'
      tag: 'QWERTY123'
      email: 'adsoft@live.com.mx'
      loginmanual: '99887766'

    xml:
      name: Operador
  Route:
    type: object
    properties:
#      id:
#        type: integer
#        format: int64
      routeid:
        type: string

      companyid:
        type: string
 
      customerid:
        type: string
      placesource:
        type: string
      placetarget:
        type: string
      
#    example:
#      codigo: 100
#      descripcion: 'viaje demo 1'
#      id: 0
#      origen: 'orizaba'
#      destino: 'ciudad mexico'
    xml:
      name: Viaje
  GetRouteListResponse:
    required:
      - routes
    properties:
      routes:
        type: array
        items:
          $ref: "#/definitions/Route"
  RouteDetail:
    type: object
    properties:
#      id:
#        type: integer
#        format: int64

      routeid:
        type: string

      routedetailid:
        type: string
 
      customerid:
        type: string

      placeid:
        type: string
      status:
        type: string
      position:
        type: integer
      
#    example:
#      codigo: 100
#      descripcion: 'viaje demo 1'
#      id: 0
#      origen: 'orizaba'
#      destino: 'ciudad mexico'
    xml:
      name: Viaje
  GetRouteDetailListResponse:
    required:
      - routes
    properties:
      routes:
        type: array
        items:
          $ref: "#/definitions/RouteDetail"
...
  Viaje:
    type: object
    properties:
#      id:
#        type: integer
#        format: int64
      orderid:
        type: integer
      driverid:
        type: string
      truckid:
        type: string
      trailerid1:
        type: string
      dollyid:
        type: string
      trailerid2:
        type: string
      zone: 
        type: string       
      assigndate: 
        type: string
      teadate: 
        type: string
      enddate: 
        type: string
      teastatus: 
        type: string
      source:
        type: string      
      sourceaddr:
        type: string      
      target: 
        type: string      
      targetaddr:
        type: string      
      km: 
        type: string      
      lt: 
        type: string      
      km_lt: 
        type: string      
      price_lt:
        type: string      
      cost:
        type: string
      routeid:
        type: string
      orderadminid:
        type: integer
      orderstatusid:
        type: integer
#    example:
#      codigo: 100
#      descripcion: 'viaje demo 1'
#      id: 0
#      origen: 'orizaba'
#      destino: 'ciudad mexico'
    xml:
      name: Viaje
  GetViajeListResponse:
    required:
      - operadores
    properties:
      operadores:
        type: array
        items:
          $ref: "#/definitions/Viaje"
  ViajeDetail:
    type: object
    properties:
      id:
        type: number
      companyid:
        type: string
      customerid:
        type: string
      orderid:
        type: number
      orderdetailid:
        type: number
      routedetailid:
        type: string
      orderdetaildescription:
        type: string
      orderdetailmessage:
        type: string      
      orderdetailarrivedate:
        type: string      
      orderdetailactivity: 
        type: string      
      orderdetailproductid:
        type: string      
      orderdetailproductdescription: 
        type: string      
      orderdetailproductquantity: 
        type: string      
      orderdetailproductunitid: 
        type: string      
      orderdetailproductunitdescription:
        type: string      
      orderdetailstatus:
        type: string
#    example:
#      codigo: 100
#      descripcion: 'viaje demo 1'
#      id: 0
#      origen: 'orizaba'
#      destino: 'ciudad mexico'
    xml:
      name: Viaje
  Order:
    required:
      - order
      - orderdetail
    properties:
    
      order:
        $ref: "#/definitions/Viaje"
      orderdetail:
        type: array
        items:
          $ref: "#/definitions/ViajeDetail"
  GetViajeDetailListResponse:
    required:
      - orderdetail
    properties:
      orderdetail:
        type: array
        items:
          $ref: "#/definitions/ViajeDetail"
...
  GetOperadorResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
      driverid: 
        type: string
        description: codigo del operador
      name:
        type: string
        description: nombre del operador
      lastname:
        type: string
      license:
        type: string
      licensetype:
        type: string
      licensedue:
        type: string
      companyid:
        type: string
      nfctag:
        type: string
      email:
        type: string
      password:
        type: string


        #  required:
  #    - id
  #    - codigo
  #    - nombre
  GetOperadoresListResponse:
    required:
      - operadores
    properties:
      operadores:
        type: array
        items:
          $ref: "#/definitions/GetOperadorResponse"
...
  ErrorResponse:
    type: object
    properties:
      error:
        $ref: '#/definitions/ErrorIntResponse'
  ErrorIntResponse:
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string
    required:
      - message